refactor: 일기 작성 페이지의 Hashtag 부분을 리팩토링하였습니다.
